Made the file architecture structure more component base and added icon for the app
